Improved code quality of BLE connectors: add missing docstrings, fixed typos and optimized code.
