+++ whad/ble/connector/injector.py
@@ -1,13 +1,22 @@
-from whad.ble.connector import BLE
-from whad.ble.exceptions import ConnectionLostException, NotSynchronized
-from whad.ble import UnsupportedCapability, message_filter, BleDirection
-from whad.hub.ble import Injected, Synchronized
-from whad.ble.sniffing import SynchronizedConnection
-from whad.ble.injecting import InjectionConfiguration
-from scapy.layers.bluetooth4LE import BTLE, BTLE_ADV, BTLE_DATA
+"""Bluetooth Low Energy packet injection module.
+
+This module provides the :class:`whad.ble.connector.injection.Injector` class
+to perform packet injection.
+"""
 from time import sleep
 
+from scapy.layers.bluetooth4LE import BTLE, BTLE_ADV, BTLE_DATA
+
+from whad.ble.connector.base import BLE
+from whad.ble.exceptions import NotSynchronized
+from whad.ble import UnsupportedCapability, message_filter
+from whad.hub.ble import Injected, Direction as BleDirection
+from whad.ble.sniffing import SynchronizedConnection
+from whad.ble.injecting import InjectionConfiguration
+
 class Injector(BLE):
+    """BLE injecion connector.
+    """
 
     def __init__(self, device, connection=None):
         super().__init__(device)
@@ -22,6 +31,8 @@ class Injector(BLE):
 
 
     def _enable_configuration(self):
+        """Enable configuration for injection.
+        """
         if self.__configuration.synchronize:
             if self.__configuration.active_connection is not None:
                 if not self.can_sniff_active_connection():
@@ -32,7 +43,8 @@ class Injector(BLE):
                 channel_map = self.__configuration.active_connection.channel_map
                 hop_interval = self.__configuration.active_connection.hop_interval
                 hop_increment = self.__configuration.active_connection.hop_increment
-                self.sniff_active_connection(access_address, crc_init, channel_map, hop_interval, hop_increment)
+                self.sniff_active_connection(access_address, crc_init, channel_map,
+                                             hop_interval, hop_increment)
             else:
                 if not self.can_sniff_new_connection():
                     raise UnsupportedCapability("Sniff")
@@ -52,17 +64,22 @@ class Injector(BLE):
             return True
         return True
 
-    def configure(self, active_connection=None, synchronize=False, channel=37, filter="FF:FF:FF:FF:FF:FF"):
+    def configure(self, active_connection=None, synchronize=False, channel=37,
+                  bd_filter="FF:FF:FF:FF:FF:FF"):
+        """Configure this connector to target an active connection.
+        """
         self.stop()
-        self.__configuration.active_connection = active_connection if active_connection is not None else None
+        self.__configuration.active_connection = active_connection
         self.__configuration.synchonize = synchronize
         self.__configuration.channel = channel
-        self.__configuration.filter = filter
+        self.__configuration.filter = bd_filter
         self._enable_configuration()
 
 
     @property
     def configuration(self):
+        """Current injection configuration.
+        """
         return self.__configuration
 
     @configuration.setter
@@ -72,18 +89,22 @@ class Injector(BLE):
         self._enable_configuration()
 
     def inject(self, packet):
+        """Inject packet in current connection.
+        """
         if self.__configuration.raw:
             return self.raw_inject(packet)
-        elif self.__configuration.inject_to_slave:
+        if self.__configuration.inject_to_slave:
             return self.inject_to_slave(packet)
-        elif self.__configuration.inject_to_master:
+        if self.__configuration.inject_to_master:
             return self.inject_to_master(packet)
-        else:
-            return False
+
+        # Cannot inject
+        return False
 
     def raw_inject(self, packet):
         """
-        Inject a raw packet directly, according to the channel provided in configuration or metadata.
+        Inject a raw packet directly, according to the channel provided in
+        configuration or metadata.
         """
         if BTLE in packet:
             access_address = packet.access_addr
@@ -94,6 +115,9 @@ class Injector(BLE):
                 access_address = self.__connection.access_address
             else:
                 access_address = 0x11223344 # default value
+        else:
+            # Fail, no access address found
+            return None
 
         if self.__configuration.channel is not None:
             channel = self.__configuration.channel
@@ -102,26 +126,36 @@ class Injector(BLE):
         else:
             channel = 37 # fallback to channel 37
 
-        return self.send_pdu(packet, access_address=access_address, conn_handle=channel, direction=BleDirection.UNKNOWN)
+        return self.send_pdu(packet, access_address=access_address, conn_handle=channel,
+                             direction=BleDirection.UNKNOWN)
 
     def inject_to_slave(self, packet):
+        """Inject packet in existing connection, targeting the advertiser.
+        """
         if self.__connection is not None:
-            self.send_pdu(packet, access_address=self.__connection.access_address, direction=BleDirection.INJECTION_TO_SLAVE)
+            self.send_pdu(packet, access_address=self.__connection.access_address,
+                          direction=BleDirection.INJECTION_TO_SLAVE)
             message = self.wait_for_message(filter=message_filter(Injected))
             return (message.success, message.injection_attempts)
-        else:
-            raise NotSynchronized()
+
+        # Not synchronized
+        raise NotSynchronized()
 
     def inject_to_master(self, packet):
+        """Inject packet in existing connection, targeting the initiator.
+        """
         if self.__connection is not None:
-            self.send_pdu(packet, access_address=self.__connection.access_address, direction=BleDirection.INJECTION_TO_MASTER)
+            self.send_pdu(packet, access_address=self.__connection.access_address,
+                          direction=BleDirection.INJECTION_TO_MASTER)
             message = self.wait_for_message(filter=message_filter(Injected))
             return (message.success, message.injection_attempts)
-        else:
-            raise NotSynchronized()
+
+        # Not synchronized
+        raise NotSynchronized()
 
 
-    def on_synchronized(self, access_address=None, crc_init=None, hop_increment=None, hop_interval=None, channel_map=None):
+    def on_synchronized(self, access_address=None, crc_init=None, hop_increment=None,
+                        hop_interval=None, channel_map=None):
         self.__connection = SynchronizedConnection(
             access_address = access_address,
             crc_init = crc_init,
@@ -131,7 +165,7 @@ class Injector(BLE):
         )
         self.__synchronized = True
 
-    def on_desynchronized(self, access_address):
+    def on_desynchronized(self, access_address=None):
         if access_address == self.__connection.access_address:
             self.__synchronized = False
             self.__connection = None
